- Fixed issues with compiling with Visual Studio 2015. - Bumped CMake requirement to 2.8.7 so we can require generator expressions and not have to repeat outselves in endif().

+++ src/g_shared/a_playerpawn.cpp
@@ -248,7 +248,7 @@ void APlayerPawn::Tick()
 	{
 		// A negative scale is used to prevent G_AddViewAngle/G_AddViewPitch
 		// from scaling with the FOV scale.
-		desired *= fabs(player->ReadyWeapon->fovscale);
+		desired *= (float)fabs(player->ReadyWeapon->fovscale);
 	}
 	if (player->FOV != desired)
 	{
